% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdtp.R
\name{print.summary.rdtp}
\alias{print.summary.rdtp}
\title{Print method for summary.rdtp objects}
\usage{
\method{print}{summary.rdtp}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"summary.rdtp"}.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for summary.rdtp objects
}
